유니티 데이터 연결 및 토큰 재발급 적용
UICamera에 데이터 저장 및 데이터 관리 스크립트 작성

토큰 재발급에 관련된 모든 스크립트 수정

친구 행성 연결을 위한 코드 일부 수정

공지사항 API 연결

+++ Assets/Scripts/UI/PopUpUI/Menu/ItemStore/UI_ItemBuy.cs
@@ -145,6 +145,11 @@ public class UI_ItemBuy : UI_PopupMenu
 
     public void BuyBtnClick(PointerEventData data)
     {
+        ExBuy();
+    }
+
+    void ExBuy()
+    {
         string[] hN = { Define.JWT_ACCESS_TOKEN,
                         "User-Id" };
         string[] hV = { Managers.Player.GetString(Define.JWT_ACCESS_TOKEN),
@@ -163,14 +168,21 @@ public class UI_ItemBuy : UI_PopupMenu
             {
                 Debug.Log(uwr.downloadHandler.text);
                 parent.GetComponent<UI_ItemStore>().SetPoint(response.result.point);
-                parent.GetComponent<UI_ItemStore>().DeleteItem(response.result.itemId);
+
+                if(response.result.maxCnt <= 0)
+                    parent.GetComponent<UI_ItemStore>().DeleteItem(response.result.itemId);
+
+                Debug.Log($"{buyAmountSlider.value}개 구매");
+                ClosePopupUI();
+            }
+            else if (response.code == 6000)
+            {
+                Managers.Player.SendTokenRequest(ExBuy);
             }
             else
             {
                 Debug.Log(response.message);
             }
-            Debug.Log($"{buyAmountSlider.value}개 구매");
-            ClosePopupUI();
         }, hN, hV);
     }
